@@ -14,32 +14,32 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.18.6 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.1
-  html: ^0.14.0+3
-  google_fonts: ^1.1.0
-  transparent_image: ^1.0.0
-  photo_view: ^0.9.2
-  autocomplete_textfield: ^1.7.3
-  path_provider: ^1.6.10
+  http: ^1.1.0
+  html: ^0.15.4
+  google_fonts: ^6.1.0
+  transparent_image: ^2.0.1
+  photo_view: ^0.14.0
+  autocomplete_textfield: ^2.0.1
+  path_provider: ^2.1.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.7.5
+  flutter_launcher_icons: ^0.13.1
 
 flutter_icons:
   android: true
   ios: true
-  image_path: "C:\\Users\\FAMILIA\\Desktop\\dart\\projetos_flutter\\leitor_de_manga\\lib\\icon\\icon.png"
+  #image_path: "C:\\Users\\FAMILIA\\Desktop\\dart\\projetos_flutter\\leitor_de_manga\\lib\\icon\\icon.png"
 
 
 # For information on the generic Dart part of this file, see the
@@ -57,8 +57,8 @@ flutter:
   # assets:
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
-  assets: 
-    - C:\Users\FAMILIA\Desktop\dart\projetos_flutter\leitor_de_manga\lib\icon\icon.png
+  #assets:
+    #- C:\Users\FAMILIA\Desktop\dart\projetos_flutter\leitor_de_manga\lib\icon\icon.png
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
 
