@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:leitor_de_manga/pages/manga_page.dart';
 import 'package:leitor_de_manga/scraper/scraper.dart';
@@ -10,14 +11,14 @@ class SearchPage extends StatefulWidget {
   final UnionScraper scraper;
 
 
-  SearchPage({this.scraper});
+  SearchPage({required this.scraper});
 
   @override
   _SearchPageState createState() => _SearchPageState();
 }
 
 class _SearchPageState extends State<SearchPage> {
-  GlobalKey key = GlobalKey<AutoCompleteTextFieldState<String>>();
+  final key = GlobalKey<AutoCompleteTextFieldState<String>>();
 
   @override
   Widget build(BuildContext context) {
@@ -25,14 +26,13 @@ class _SearchPageState extends State<SearchPage> {
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: AppBar(
-        brightness: Brightness.dark, // Status Bar Color Icons
         backgroundColor: Color(0xFF050B18),
         leading: IconButton(
           icon: Icon(
             Icons.arrow_back_ios,
             size: 30.0,
             color: Colors.white,
-          ), 
+          ),
           onPressed: (){
             Navigator.pop(context);
           },
@@ -72,7 +72,7 @@ class _SearchPageState extends State<SearchPage> {
               ),
               itemSubmitted: (item){
                 Navigator.push(
-                  context, 
+                  context,
                   MaterialPageRoute(
                     builder: (context) => MangaPage(
                       scraper: widget.scraper,
@@ -80,9 +80,9 @@ class _SearchPageState extends State<SearchPage> {
                     ),
                   ),
                 );
-              }, 
-              key: key, 
-              suggestions: Utils.allTitles, 
+              },
+              key: key,
+              suggestions: Utils.allTitles,
               itemBuilder: (context, item){
                 return Card(
                   elevation: 3.0,
@@ -97,16 +97,16 @@ class _SearchPageState extends State<SearchPage> {
                     ),
                   ),
                 );
-              }, 
+              },
               itemSorter: (a,b){
                 return a.compareTo(b);
-              }, 
+              },
               itemFilter: (item, query){
                 return item.toLowerCase().startsWith(query.toLowerCase());
               }
             ),
           ),
-        ),
+        ), systemOverlayStyle: SystemUiOverlayStyle.light,
       ),
     );
   }
